+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * new `Konva.Transformer` group that allow simple resize, and rotate of a shape.
+* Add ability to remove event by callback `node.off('event', callback)`
 
 ## Changed